# Labor.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Labor.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Address": {
      "get": {
        "tags": [
          "Address"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AddressIEnumerableApiResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Address"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddressDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Int32ApiResponse"
            }
          }
        }
      }
    },
    "/api/Address/{addressId}": {
      "get": {
        "tags": [
          "Address"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AddressApiResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Address"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AddressDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Address"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/api/Address/{addressId}/set-default": {
      "put": {
        "tags": [
          "Address"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "addressId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Address": {
      "required": [
        "addressType",
        "city",
        "country",
        "state",
        "street",
        "zipCode"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "addressType": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "street": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "city": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "state": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "country": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "zipCode": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        },
        "isDefault": {
          "type": "boolean"
        },
        "isActive": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        }
      },
      "additionalProperties": false
    },
    "AddressApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/Address"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
    "AddressDto": {
      "required": [
        "city",
        "country",
        "state",
        "street",
        "zipCode"
      ],
      "type": "object",
      "properties": {
        "addressId": {
          "format": "int32",
          "type": "integer"
        },
        "addressType": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "street": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "city": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "state": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "country": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "zipCode": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        },
        "isDefault": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "AddressIEnumerableApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": { },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Cart": {
      "required": [
        "dailyRate",
        "requiredHours",
        "totalAmount"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "employerId": {
          "format": "int32",
          "type": "integer"
        },
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "requiredHours": {
          "format": "int32",
          "type": "integer"
        },
        "dailyRate": {
          "format": "double",
          "type": "number"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "workDescription": {
          "maxLength": 1000,
          "minLength": 0,
          "type": "string"
        },
        "preferredDate": {
          "format": "date-time",
          "type": "string"
        },
        "is_deleted": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "employer": {
          "$ref": "#/definitions/User"
        },
        "labor": {
          "$ref": "#/definitions/Laborer"
        }
      },
      "additionalProperties": false
    },
...
    "Int32ApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "format": "int32",
          "type": "integer"
        },
        "errors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "additionalProperties": false
    },
...
    "LaborReview": {
      "required": [
        "rating"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "orderItemId": {
          "format": "int32",
          "type": "integer"
        },
        "employerId": {
          "format": "int32",
          "type": "integer"
        },
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "int32",
          "maximum": 5,
          "minimum": 1,
          "type": "integer"
        },
        "comment": {
          "maxLength": 1000,
          "minLength": 0,
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "orderItem": {
          "$ref": "#/definitions/OrderItem"
        },
        "employer": {
          "$ref": "#/definitions/User"
        },
        "labor": {
          "$ref": "#/definitions/Laborer"
        }
      },
      "additionalProperties": false
    },
...
    "LaborSkill": {
      "required": [
        "proficiencyLevel",
        "skillName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "skillName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "proficiencyLevel": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "labor": {
          "$ref": "#/definitions/Laborer"
        }
      },
      "additionalProperties": false
    },
...
    "LaborType": {
      "required": [
        "typeName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "typeName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "dailyRate": {
          "format": "double",
          "type": "number"
        },
        "isActive": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedBy": {
          "type": "string"
        },
        "labors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Laborer"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Laborer": {
      "required": [
        "dailyRate"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "laborTypeId": {
          "format": "int32",
          "type": "integer"
        },
        "specialization": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "experienceYears": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "totalReviews": {
          "format": "int32",
          "type": "integer"
        },
        "dailyRate": {
          "format": "double",
          "type": "number"
        },
        "minimumHours": {
          "format": "int32",
          "type": "integer"
        },
        "maximumHours": {
          "format": "int32",
          "type": "integer"
        },
        "availabilityStatus": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "isVerified": {
          "type": "boolean"
        },
        "isActive": {
          "type": "boolean"
        },
        "is_deleted": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedBy": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "laborType": {
          "$ref": "#/definitions/LaborType"
        },
        "laborSkills": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborSkill"
          }
        },
        "carts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Cart"
          }
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        },
        "laborReviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborReview"
          }
        }
      },
      "additionalProperties": false
    },
...
    "Order": {
      "required": [
        "orderNumber",
        "totalAmount"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "orderNumber": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "employerId": {
          "format": "int32",
          "type": "integer"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "orderStatus": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "paymentStatus": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "workAddressId": {
          "format": "int32",
          "type": "integer"
        },
        "scheduledDate": {
          "format": "date-time",
          "type": "string"
        },
        "completedDate": {
          "format": "date-time",
          "type": "string"
        },
        "cancelledDate": {
          "format": "date-time",
          "type": "string"
        },
        "is_deleted": {
          "type": "boolean"
        },
        "cancelReason": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "employer": {
          "$ref": "#/definitions/User"
        },
        "workAddress": {
          "$ref": "#/definitions/Address"
        },
        "orderItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        },
        "orderTrackings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderTracking"
          }
        }
      },
      "additionalProperties": false
    },
...
    "OrderItem": {
      "required": [
        "dailyRate",
        "requiredHours",
        "totalAmount"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "laborId": {
          "format": "int32",
          "type": "integer"
        },
        "requiredHours": {
          "format": "int32",
          "type": "integer"
        },
        "dailyRate": {
          "format": "double",
          "type": "number"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "workDescription": {
          "maxLength": 1000,
          "minLength": 0,
          "type": "string"
        },
        "itemStatus": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "actualHours": {
          "format": "int32",
          "type": "integer"
        },
        "startTime": {
          "format": "date-time",
          "type": "string"
        },
        "endTime": {
          "format": "date-time",
          "type": "string"
        },
        "order": {
          "$ref": "#/definitions/Order"
        },
        "labor": {
          "$ref": "#/definitions/Laborer"
        },
        "laborReviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborReview"
          }
        }
      },
      "additionalProperties": false
    },
...
    "OrderTracking": {
      "required": [
        "status"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "location": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "format": "uuid",
          "type": "string"
        },
        "order": {
          "$ref": "#/definitions/Order"
        }
      },
      "additionalProperties": false
    },
...
    "Person": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "profilePicture": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "format": "int32",
          "type": "integer"
        },
        "updatedBy": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "Role": {
      "required": [
        "roleName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "roleName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedBy": {
          "type": "string"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        },
        "rolePermissions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RolePermission"
          }
        }
      },
      "additionalProperties": false
    },
    "RolePermission": {
      "required": [
        "featureName"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "featureName": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "canView": {
          "type": "boolean"
        },
        "canCreate": {
          "type": "boolean"
        },
        "canEdit": {
          "type": "boolean"
        },
        "canDelete": {
          "type": "boolean"
        },
        "isActive": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedBy": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/Role"
        }
      },
      "additionalProperties": false
    },
...
    "User": {
      "required": [
        "mobileNumber"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "mobileNumber": {
          "maxLength": 15,
          "minLength": 0,
          "type": "string"
        },
        "passwordHash": {
          "maxLength": 255,
          "minLength": 0,
          "type": "string"
        },
        "personId": {
          "format": "int32",
          "type": "integer"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "isTemporaryPassword": {
          "type": "boolean"
        },
        "isActive": {
          "type": "boolean"
        },
        "isMobileVerified": {
          "type": "boolean"
        },
        "isEmailVerified": {
          "type": "boolean"
        },
        "isProfileComplete": {
          "type": "boolean"
        },
        "is_deleted": {
          "type": "boolean"
        },
        "lastLoginAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedBy": {
          "type": "string"
        },
        "person": {
          "$ref": "#/definitions/Person"
        },
        "role": {
          "$ref": "#/definitions/Role"
        },
        "addresses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Address"
          }
        },
        "laborProfile": {
          "$ref": "#/definitions/Laborer"
        },
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        },
        "carts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Cart"
          }
        },
        "laborReviews": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/LaborReview"
          }
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}